useorocos: extended DESTDIR support for the case where multiple packages are built in the same cmake workspace
This patch extends and simplifies the DESTDIR support introduced in 85fc9d3f0bbb6fe31c81b97b7df8c2a49160dfb0.
It should have no effect on normal builds where the DESTDIR environment variable is not set.

The original patch assumed that DESTDIR is already set during the configure step, like with

```
DESTDIR=/tmp/destdir catkin_make_isolated --install ...
```

In a pure catkin workspace packages can be built for a whole workspace with a standard debian/rules file
and debhelper's support for the cmake build system. In this case DESTDIR is only set during the install step
and the previous patch was not effective, causing RPATH-related errors during runtime.

The INSTALL_RPATH is still constructed based on the information from the pkg-config files for packages in
*overlayed* workspaces or system packages. Additionally, libraries built in the *same* workspace will now export
their library dirs in the *install-space* via the cmake cache so that other targets that link to it can already
set their RPATH correctly.

In order to make sure that no paths within DESTDIR end up in the installed RPATH of any library, a new cmake macro
`orocos_set_install_rpath(target rpath)` has been introduced that cleans up all entries in the rpath list and
removes duplicates. This could be an issue if RTT itself is installed in DESTDIR together with other packages, as
`OROCOS-RTT_LIBRARY_DIRS` would have the DESTDIR prefix.

Signed-off-by: Johannes Meyer <johannes@intermodalics.eu>