Timer: fixed wake up time on 32 bit systems
The existing comment is right:
We can't use infinite as the OS may internally use time_spec, which can not
represent as much in the future (until 2038) // XXX Year-2038 Bug

but used value of InfiniteNSecs/4 exceeds max in32 seconds.

Changed to use value to be converted within max int32 seconds.