Added some missing visibility attributes if RTT is compiled with default visiblity hidden
Note that by default all symbols are visible in GCC unless they are explicitly declared as hidden.
This patch does not change this. For some reason the C functions defined in tlsf.c always end up
as private symbols in the RTT core library and lead to linker errors later on.

Signed-off-by: Johannes Meyer <johannes@intermodalics.eu>