tests: fixed dev_test for gcc 5
DigitalIniput expects a const bool & in its constructor and stores it in
a member variable. The dev_test passed a boolean literal, but C++ creates a temporary
in this case which is only valid during the call.

At least with gcc 5 the dev_test consistenly failed because of this issue:
/home/travis/build/orocos-toolchain/rtt/tests/dev_test.cpp(56): error in "testClasses": check din.isOn() failed

Signed-off-by: Johannes Meyer <johannes@intermodalics.eu>