Added virtual method RTT::os::ThreadInterface::getCpuAffinity() and fixed comparison in RTT::TimerThread
The CPU affinity comparison in RTT::TimerThread::Instance() added in 2cfecd7 was not taking into account
which CPUs are actually available and therefore created a new timer thread in almost all cases
on target gnulinux. This patch makes getCpuAffinity() also available for RTT::os::MainThread and uses the
main thread's CPU affinity as a default value for TimerThread instances.

Signed-off-by: Johannes Meyer <johannes@intermodalics.eu>