deployment: revert API changes and fixed ComponentLoader::unloadComponent() for aliased components
https://github.com/orocos-toolchain/rtt/pull/66 introduced aliases for CORBA component servers
and added an optional name argument to ComponentLoader::unloadComponent(tc). This change has
been reverted, because it could lead to inconsistencies in the internal component map, if a
TaskContext is destroyed based on the given name but with an unrelated instance pointer.

The new approach should be safer as it does not rely on the name or alias of the instance to
be unloaded at all.

Signed-off-by: Johannes Meyer <johannes@intermodalics.eu>