tests: unified result of independent merges of state_test.cpp compared to upstream
The independent resolution of conflicts during several merge commits and cherry-picks
have caused some small differences in the state_test.cpp implementation compared to
upstream. This commit unifies the two implementations, while excluding changes in
one of the two branches which have not yet been applied to the other.

The only difference other than instrumentation, logging and variable renamings is
the addition of two yields in testStateEvents which have been introduced in
9e20c1a516b0d9ca443e3629203327d5a000e549 but deleted during the merge in
8c5d80da23c602ca9a07cf3ddf9419a46d6aa3d0.

Signed-off-by: Johannes Meyer <johannes@intermodalics.eu>