base/tests: fix a potential conflicting read in DataObjectLockFree<T>
The write_ptr was incremented in a loop in DataObjectLockFree<T>::Set() and it is not
guaranteed that this section cannot be executed by two concurrent writers if one of them
already modified the write_ptr at least once before the other starts to write. This situation
can be prevented by first modifying a copy of the write_ptr and only setting the shared
write_ptr once.

In order to trigger this case without the patch in DataObjectLockFree.hpp an option to sleep
during the assignment of the sample has been added to the dataflow_performance_test.cpp,
independent for reading and writing. Different from the delay caused by copying large samples
the sleep command also triggers the scheduler and more likely triggers failure mode where one
of thread is preempted during the assignment and while still holding a (non-exclusive) read
or write lock.

Thanks to @kevindm for helping debugging this.

Signed-off-by: Johannes Meyer <johannes@intermodalics.eu>